[chore #162163792] spilts validation.js into two
